requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2.2
psycopg2-binary>=2.9
Pillow>=9.0
django-cors-headers>=4.0
